Share ResolutionResult builder test fixtures and test concurrency deterministically
Extract the graph-building fixtures shared by StreamingResolutionResultBuilderTest
and the concurrency test into AbstractResolutionResultBuilderTest.

Rework the concurrency test to force the exact lock-ordering interleaving with
latches instead of hammering the race: it holds the graph monitor, parks a
variants reader on it (while that reader holds the component monitor), then reads
the same component's dependents from the thread holding the graph monitor. A
deadlock is reported immediately by ThreadMXBean.findDeadlockedThreads(), so the
test needs no timeout. Verified to deadlock against the pre-fix code and to pass
with the fix.
